# Conflicts:
#	game_dx12/x64/Debug/game_dx12.exe
#	game_dx12/x64/Debug/game_dx12.ilk
#	game_dx12/x64/Debug/game_dx12.pdb
#	game_dx12/x64/Release/game_dx12.exe
#	game_dx12/x64/Release/game_dx12.iobj
#	game_dx12/x64/Release/game_dx12.ipdb
#	game_dx12/x64/Release/game_dx12.pdb

+++ game_dx12/game_dx12/inc/KDL_Dx12/Sprite.h
@@ -167,7 +167,7 @@ namespace KDL
 				bool z_test = false);
 			static size_t GetTextWidth(const std::string& text, float size_w, float scale_w, float space)
 			{
-				return text.size() * (size_w + space) * scale_w;
+				return static_cast<size_t>(text.size() * (size_w + space) * scale_w);
 			}
 			template <class T = FLOAT>
 			Number2<T> GetSize() { return { static_cast<T>(m_tex_desc.Width), static_cast<T>(m_tex_desc.Height) }; }
